+++ src/effects/spells/hypnoticPattern.js
@@ -5,5 +5,9 @@ export function hypnoticPatternEffect(document) {
   effectHypnoticPatternCharmed.changes.push(generateStatusEffectChange("Charmed"));
   document.effects.push(effectHypnoticPatternCharmed);
 
+  let effectHypnoticPatternIncapacitated = baseSpellEffect(document, document.name);
+  effectHypnoticPatternIncapacitated.changes.push(generateStatusEffectChange("Incapacitated"));
+  document.effects.push(effectHypnoticPatternIncapacitated);
+
   return document;
 }