+++ CHANGELOG.md
@@ -1,10 +1,11 @@
-# NEXT
+# 3.3.10
 
 * Slayer's Prey effect.
 * Hypnotic Pattern effect improvements.
 * Green Flame Blade spell effect tweaks.
 * Ensnaring Strike effect by Elwin#1410. Thanks Elwin!
 * Added `.api` extension to `game.modules.get("ddb-importer")` to expose functions exposed through the window previously (old method will still work indefinitely).
+* Added homebrew filter for item and spell munching.
 
 # 3.3.9